refactor(oscript): вынести общую обвязку autumn-индексов в AbstractAutumnLibraryIndex
AutumnBeanIndex, AutumnInjectionPointIndex и AutumnCollectionIndex дублировали
одинаковую обвязку ленивой сборки: барьер ready/ensureBuilt, @EventListener'ы
(invalidate по переиндексации библиотек, точечная переиндексация на правку,
удаление вклада документа), rebuild + indexDocument, isAnnotationDefinition.
Sonar справедливо ругался на дублирование нового кода (new injection-point индекс
повторял bean-индекс).

Выносим общую часть в abstract AbstractAutumnLibraryIndex; наследники реализуют
clearIndex/indexClass/removeByUri. Поведение не меняется (Spring наследует
@EventListener в подклассах-бинах), что подтверждают модульные и интеграционные
тесты.
